Create delete and Update methods for RateClimbModal
Modal changes depending on if user has previously rated climb.
Checks with `saveRating`, which checks user against previous.

If they have already rated, button is updated and modal shows
update modal, which has Delete button

Delete button simply fetches, and throws and error and it doesn't exist.

Used onSubmit to dynicially adjust request according to `saveRating`

Tested, and tested again.

Different messages are displayed depending on 201 or 204 status
and urlExt will pass in rating id if it exists.

Either PATCH or POST are sent depending on `saveRating` as well

I would like to refactor, but for now it is working